services/content_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
drf-spectacular>=0.27
django-filter>=24.2
dj-database-url>=2.1
psycopg2-binary>=2.9
django-redis>=5.4
redis>=5.0
celery>=5.3
django-storages>=1.14
boto3>=1.34
pytest>=7.4
pytest-django>=4.8


